Ignore EOF error when unmarshaling submit_sm_resp with no message_id (#119)
Fix for #117 and #118 

I don't fully understand the reason behind #113 , but that change is
causing issues when submit_sm_resp returns with command_status != 0. I
propose we ignore the io.EOF error when unmarshalling the message_id.

I also added a test to identify this issue in future.

Reminder, as per the spec: The submit_sm_resp PDU Body is not returned
if the command_status field contains a non-zero value.